+++ packages/smarthr-ui/src/components/FormGroup/constants.ts
@@ -1,0 +1,2 @@
+export const CHILDREN_WRAPPER_INPUT_SELECTOR = `.smarthr-ui-FormControl-childrenWrapper [data-smarthr-ui-input="true"]`
+export const LABEL_TEXT_SELECTOR = '.smarthr-ui-FormControl-labelText'
